Move split resize regression to package tests
